requirements.txt
autopep8==1.5.7
bcrypt==3.1.4
blinker==1.4
certifi==2016.2.28
cffi==1.14.6
click==6.7
Flask==1.0
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.2
greenlet==1.1.1
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==2.0.1
Pillow==8.3.1
pycodestyle==2.7.0
pycparser==2.18
python-dotenv==0.19.0
six==1.11.0
SQLAlchemy==1.4.22
toml==0.10.2
Werkzeug==0.14.1
WTForms==2.1
